Added functionality to select a product, it is added to a list and the total price is totted up, which is used later.
All that is needed now is a query to grab all the necessary tables, the date needs to be figured out, but the employee ID is already a set variable, so that is easy to use.

Order of operations is to first:
Create new record in orders, add the current employee logged in as employeeID, the date is the current day rn, payment type is filled in by the box and the cost is filled in by the decimal priceTotal.

Orderline now needs to be created. Orderline is automatic and may be removed and a composite key may be used instead. But the order ID will just be the most recent order, and product ID will be the first thing in the list, then a new record is created with the second product in the list, then so on and so forth